+++ health.go
@@ -12,15 +12,6 @@ import (
 	"go.opencensus.io/stats"
 	"go.opencensus.io/stats/view"
 	"go.opencensus.io/tag"
-
-	"github.com/AppsFlyer/go-sundheit/checks"
-)
-
-const (
-	maxExpectedChecks = 16
-	initialResultMsg  = "didn't run yet"
-	// ValAllChecks is the value used for the check tags when tagging all tests
-	ValAllChecks = "all_checks"
 )
 
 var (
@@ -84,32 +75,6 @@ type Health interface {
 	WithCheckListener(listener CheckListener)
 }
 
-// CheckListener can be used to gain check stats or log check transitions.
-// Implementations of this interface **must not block!**
-// If an implementation blocks, it may result in delayed execution of other health checks down the line.
-// It's OK to log in the implementation and it's OK to add metrics, but it's not OK to run anything that
-// takes long time to complete such as network IO etc.
-type CheckListener interface {
-	// OnCheckStarted is called when a check with the specified name has started
-	OnCheckStarted(name string)
-
-	// OnCheckCompleted is called when the check with the specified name has completed it's execution.
-	// The results are passed as an argument
-	OnCheckCompleted(name string, result Result)
-}
-
-// Config defines a health Check and it's scheduling timing requirements.
-type Config struct {
-	// Check is the health Check to be scheduled for execution.
-	Check checks.Check
-	// ExecutionPeriod is the period between successive executions.
-	ExecutionPeriod time.Duration
-	// InitialDelay is the time to delay first execution; defaults to zero.
-	InitialDelay time.Duration
-	// InitiallyPassing indicates when true, the check will be treated as passing before the first run; defaults to false
-	InitiallyPassing bool
-}
-
 // New returns a new Health instance.
 func New() Health {
 	return &health{
@@ -154,26 +119,6 @@ func (h *health) createCheckTask(cfg *Config) *checkTask {
 	h.checkTasks[cfg.Check.Name()] = task
 
 	return &task
-}
-
-type checkTask struct {
-	stopChan chan bool
-	ticker   *time.Ticker
-	check    checks.Check
-}
-
-func (t *checkTask) stop() {
-	if t.ticker != nil {
-		t.ticker.Stop()
-	}
-}
-
-func (t *checkTask) execute() (details interface{}, duration time.Duration, err error) {
-	startTime := time.Now()
-	details, err = t.check.Execute()
-	duration = time.Since(startTime)
-
-	return
 }
 
 func (h *health) stopCheckTask(name string) {
@@ -265,16 +210,6 @@ func (h *health) IsHealthy() (healthy bool) {
 	return allHealthy(h.results)
 }
 
-func allHealthy(results map[string]Result) (healthy bool) {
-	for _, v := range results {
-		if !v.IsHealthy() {
-			return false
-		}
-	}
-
-	return true
-}
-
 func (h *health) updateResult(
 	name string, details interface{}, checkDuration time.Duration, err error, t time.Time) (result Result) {
 
@@ -336,32 +271,6 @@ func (h *health) WithCheckListener(listener CheckListener) {
 	}
 }
 
-// Result represents the output of a health check execution.
-type Result struct {
-	// the details of task Result - may be nil
-	Details interface{} `json:"message,omitempty"`
-	// the error returned from a failed health check - nil when successful
-	Error error `json:"error,omitempty"`
-	// the time of the last health check
-	Timestamp time.Time `json:"timestamp"`
-	// the execution duration of the last check
-	Duration time.Duration `json:"duration,omitempty"`
-	// the number of failures that occurred in a row
-	ContiguousFailures int64 `json:"contiguousFailures"`
-	// the time of the initial transitional failure
-	TimeOfFirstFailure *time.Time `json:"timeOfFirstFailure"`
-}
-
-// IsHealthy returns true iff the check result snapshot was a success
-func (r Result) IsHealthy() bool {
-	return r.Error == nil
-}
-
-func (r Result) String() string {
-	return fmt.Sprintf("Result{details: %s, err: %s, time: %s, contiguousFailures: %d, timeOfFirstFailure:%s}",
-		r.Details, r.Error, r.Timestamp, r.ContiguousFailures, r.TimeOfFirstFailure)
-}
-
 type status bool
 
 func (s status) asInt64() int64 {
@@ -370,38 +279,3 @@ func (s status) asInt64() int64 {
 	}
 	return 0
 }
-
-type marshalableError struct {
-	Message string `json:"message,omitempty"`
-	Cause   error  `json:"cause,omitempty"`
-}
-
-func newMarshalableError(err error) error {
-	if err == nil {
-		return nil
-	}
-
-	mr := &marshalableError{
-		Message: err.Error(),
-	}
-
-	cause := errors.Cause(err)
-	if cause != err {
-		mr.Cause = newMarshalableError(cause)
-	}
-
-	return mr
-}
-
-func (e *marshalableError) Error() string {
-	return e.Message
-}
-
-type noopCheckListener struct{}
-
-func (noop noopCheckListener) OnCheckStarted(_ string) {}
-
-func (noop noopCheckListener) OnCheckCompleted(_ string, _ Result) {}
-
-// make sure noopCheckListener implements the CheckListener interface
-var _ CheckListener = noopCheckListener{}
